+++ examples/index.js
@@ -76,110 +76,110 @@ const { account } = remme;
   console.log("batchId:", batchStatus);
   const peers = await remme.blockchainInfo.getPeers();
   console.log("peers:", peers);
-  // const receipts = await remme.blockchainInfo.getReceipts([transactions.data[1].header_signature, "9d2dc2ab673d028bc1dd8b5be8d2d885e4383a827cd0261f58334252bf807c08113207eabbd12d0786d6bba5378a791129f9c520c17597b5504d4b547ef57491"]);
-  // console.log("receipts:", receipts);
+  const receipts = await remme.blockchainInfo.getReceipts([transactions.data[1].header_signature, "9d2dc2ab673d028bc1dd8b5be8d2d885e4383a827cd0261f58334252bf807c08113207eabbd12d0786d6bba5378a791129f9c520c17597b5504d4b547ef57491"]);
+  console.log("receipts:", receipts);
 
   // Certificates Operations
-  const certificateTransactionResult = await remme.certificate.createAndStore({
-    commonName: "userName",
-    email: "user@email.com",
-    name: "John",
-    surname: "Smith",
-    countryName: "US",
-    validity: 360,
-    serial: `${Date.now()}`
-  }); // 4
-
-  const certificateTransactionCallback = async (err, response) => {
-    if (err) {
-      console.log(err);
-    }
-    console.log("store", response);
-    if (response.status === "COMMITTED") {
-      const status = await remme.certificate.check(certificateTransactionResult.certificate);
-      console.log('status:', status);
-      const revoke = await remme.certificate.revoke(certificateTransactionResult.certificate);
-      revoke.connectToWebSocket(async (err, response) => {
-        console.log("error", err);
-        console.log("revoke", response);
-        const status = await remme.certificate.check(certificateTransactionResult.certificate);
-        console.log(status);
-      })
-    }
-  }; // 7
-
-  certificateTransactionResult.connectToWebSocket(certificateTransactionCallback); // 5
-
-  remme.events.subscribe({
-    events: RemmeEvents.SwapInit,
-    // lastKnownBlockId: "db19f0e3b3f001670bebc814e238df48cef059f3f0668f57702ba9ff0c4b8ec45c7298f08b4c2fa67602da27a84b3df5dc78ce0f7774b3d3ae094caeeb9cbc82"
-  }, (err, res) => {
-    if (err) {
-      console.log(err);
-      return;
-    }
-    console.log("events", res);
-  });
-
-  const swapId = "033102e41346242476b15a3a7966eb5249271025fc7fb0b37ed3fdb4bcce3873";
-  const secret = "secretkey";
-  const secretKey = "039eaa877ff63694f8f09c8034403f8b5165a7418812a642396d5d539f90b170";
-  const secretLock = "b605112c2d7489034bbd7beab083fb65ba02af787786bb5e3d99bb26709f4f68";
-
-  const init = await remme.swap.init({
-    receiverAddress: "112007484def48e1c6b77cf784aeabcac51222e48ae14f3821697f4040247ba01558b1",
-    senderAddressNonLocal: "0xe6ca0e7c974f06471759e9a05d18b538c5ced11e",
-    amount: 100,
-    swapId,
-    createdAt: Math.floor(Date.now() / 1000)
-  });
-
-  init.connectToWebSocket(async (err, data) => {
-    if (err) {
-      console.log("err init", err);
-    } else if (data.status === "COMMITTED") {
-      console.log("data init", data);
-      const res = await remme.swap.getInfo(swapId);
-      console.log("after init info", res);
-      const pubkey = await remme.swap.getPublicKey();
-      console.log(pubkey);
-      init.closeWebSocket();
-      const setSecretLock = await remme.swap.setSecretLock(swapId, secretLock);
-      setSecretLock.connectToWebSocket(async (err, data) => {
-        if (err) {
-          console.log("set secret lock err", err);
-        }
-        console.log("data set secret lock", data);
-        if (data.status === "COMMITTED") {
-          const res = await remme.swap.getInfo(swapId);
-          console.log("after set secret lock info", res);
-          setSecretLock.closeWebSocket();
-          const approve = await remme.swap.approve(swapId);
-          approve.connectToWebSocket(async (err, data) => {
-            if (err) {
-              console.log("approve err", err);
-            }
-            console.log("data approve", data);
-            if (data.status === "COMMITTED") {
-              const res = await remme.swap.getInfo(swapId);
-              console.log("after approve", res);
-              approve.closeWebSocket();
-              const close = await remme.swap.close(swapId, secretKey);
-              close.connectToWebSocket(async (err, data) => {
-                if (err) {
-                  console.log("err close", err);
-                }
-                console.log("data close", data);
-                if (data.status === "COMMITTED") {
-                  const res = await remme.swap.getInfo(swapId);
-                  console.log("after close info", res);
-                  close.closeWebSocket();
-                }
-              });
-            }
-          });
-        }
-      });
-    }
-  });
+  // const certificateTransactionResult = await remme.certificate.createAndStore({
+  //   commonName: "userName",
+  //   email: "user@email.com",
+  //   name: "John",
+  //   surname: "Smith",
+  //   countryName: "US",
+  //   validity: 360,
+  //   serial: `${Date.now()}`
+  // }); // 4
+  //
+  // const certificateTransactionCallback = async (err, response) => {
+  //   if (err) {
+  //     console.log(err);
+  //   }
+  //   console.log("store", response);
+  //   if (response.status === "COMMITTED") {
+  //     const status = await remme.certificate.check(certificateTransactionResult.certificate);
+  //     console.log('status:', status);
+  //     const revoke = await remme.certificate.revoke(certificateTransactionResult.certificate);
+  //     revoke.connectToWebSocket(async (err, response) => {
+  //       console.log("error", err);
+  //       console.log("revoke", response);
+  //       const status = await remme.certificate.check(certificateTransactionResult.certificate);
+  //       console.log(status);
+  //     })
+  //   }
+  // }; // 7
+  //
+  // certificateTransactionResult.connectToWebSocket(certificateTransactionCallback); // 5
+  //
+  // remme.events.subscribe({
+  //   events: RemmeEvents.SwapInit,
+  //   // lastKnownBlockId: "db19f0e3b3f001670bebc814e238df48cef059f3f0668f57702ba9ff0c4b8ec45c7298f08b4c2fa67602da27a84b3df5dc78ce0f7774b3d3ae094caeeb9cbc82"
+  // }, (err, res) => {
+  //   if (err) {
+  //     console.log(err);
+  //     return;
+  //   }
+  //   console.log("events", res);
+  // });
+  //
+  // const swapId = "033102e41346242476b15a3a7966eb5249271025fc7fb0b37ed3fdb4bcce3873";
+  // const secret = "secretkey";
+  // const secretKey = "039eaa877ff63694f8f09c8034403f8b5165a7418812a642396d5d539f90b170";
+  // const secretLock = "b605112c2d7489034bbd7beab083fb65ba02af787786bb5e3d99bb26709f4f68";
+  //
+  // const init = await remme.swap.init({
+  //   receiverAddress: "112007484def48e1c6b77cf784aeabcac51222e48ae14f3821697f4040247ba01558b1",
+  //   senderAddressNonLocal: "0xe6ca0e7c974f06471759e9a05d18b538c5ced11e",
+  //   amount: 100,
+  //   swapId,
+  //   createdAt: Math.floor(Date.now() / 1000)
+  // });
+  //
+  // init.connectToWebSocket(async (err, data) => {
+  //   if (err) {
+  //     console.log("err init", err);
+  //   } else if (data.status === "COMMITTED") {
+  //     console.log("data init", data);
+  //     const res = await remme.swap.getInfo(swapId);
+  //     console.log("after init info", res);
+  //     const pubkey = await remme.swap.getPublicKey();
+  //     console.log(pubkey);
+  //     init.closeWebSocket();
+  //     const setSecretLock = await remme.swap.setSecretLock(swapId, secretLock);
+  //     setSecretLock.connectToWebSocket(async (err, data) => {
+  //       if (err) {
+  //         console.log("set secret lock err", err);
+  //       }
+  //       console.log("data set secret lock", data);
+  //       if (data.status === "COMMITTED") {
+  //         const res = await remme.swap.getInfo(swapId);
+  //         console.log("after set secret lock info", res);
+  //         setSecretLock.closeWebSocket();
+  //         const approve = await remme.swap.approve(swapId);
+  //         approve.connectToWebSocket(async (err, data) => {
+  //           if (err) {
+  //             console.log("approve err", err);
+  //           }
+  //           console.log("data approve", data);
+  //           if (data.status === "COMMITTED") {
+  //             const res = await remme.swap.getInfo(swapId);
+  //             console.log("after approve", res);
+  //             approve.closeWebSocket();
+  //             const close = await remme.swap.close(swapId, secretKey);
+  //             close.connectToWebSocket(async (err, data) => {
+  //               if (err) {
+  //                 console.log("err close", err);
+  //               }
+  //               console.log("data close", data);
+  //               if (data.status === "COMMITTED") {
+  //                 const res = await remme.swap.getInfo(swapId);
+  //                 console.log("after close info", res);
+  //                 close.closeWebSocket();
+  //               }
+  //             });
+  //           }
+  //         });
+  //       }
+  //     });
+  //   }
+  // });
 })();