+++ packages/remme-transaction-service/src/index.ts
@@ -1,12 +1,13 @@
 import { RemmeMethods, IRemmeRest } from "remme-rest";
 import { IRemmeAccount } from "remme-account";
-import { createHash } from "crypto";
+import { sha512 } from "remme-utils";
 import * as protobuf from "sawtooth-sdk/protobuf";
 
 import { IRemmeTransactionService } from "./interface";
 import { TransactionCreatePayload, BaseTransactionResponse, IBaseTransactionResponse } from "./models";
 
 class RemmeTransactionService implements IRemmeTransactionService {
+
     // index signature
     [key: string]: any;
 
@@ -27,20 +28,17 @@ class RemmeTransactionService implements IRemmeTransactionService {
             payloadBytes,
         } = settings;
 
-        // const { pubkey: batcherPublicKey } = await this._remmeRest
-        //     .getRequest<{pubkey: string}>(RemmeMethods.nodeKey);
-        const batcherPublicKey = await this._remmeRest
-            .sendRequest<string>(RemmeMethods.nodeKey);
+        const batcherPublicKey = await this._remmeRest.sendRequest<string>(RemmeMethods.nodeKey);
+
         const transactionHeaderBytes = protobuf.TransactionHeader.encode({
             familyName,
             familyVersion,
             inputs: [ ...inputs, this._remmeAccount.address ],
             outputs: [ ...outputs, this._remmeAccount.address ],
             signerPublicKey: this._remmeAccount.publicKeyHex,
-            nonce: this.getNonce(),
+            nonce: sha512(Math.floor(Math.random() * 1000).toString()),
             batcherPublicKey,
-            // payloadSha512: this.generateHash(payloadBytes.toString()),
-            payloadSha512: createHash("sha512").update(new Buffer(payloadBytes)).digest("hex"),
+            payloadSha512: sha512(payloadBytes),
         }).finish();
 
         const signature = this._remmeAccount.sign(transactionHeaderBytes);
@@ -61,14 +59,6 @@ class RemmeTransactionService implements IRemmeTransactionService {
     }
 
     public async send(transaction: string): Promise<IBaseTransactionResponse> {
-        // const apiResult = await this._remmeRest
-        //     .postRequest<{transaction: string}, {batch_id: string, error?: string}>
-        //     (RemmeMethods.transaction, { transaction });
-        // return new BaseTransactionResponse(
-        //     this._remmeRest.nodeAddress(),
-        //     this._remmeRest.sslMode(),
-        //     apiResult.batch_id,
-        // );
         const batchId = await this._remmeRest
             .sendRequest<{data: string}, string>
             (RemmeMethods.transaction, { data: transaction });
@@ -79,15 +69,6 @@ class RemmeTransactionService implements IRemmeTransactionService {
         );
     }
 
-    // public generateHash(data: string): string {
-    //     const certSHA512 = forge.md.sha512.create().update(data);
-    //     return certSHA512.digest().toHex();
-    // }
-
-    private getNonce(): string {
-        // return this.generateHash(Math.floor(Math.random() * 1000).toString());
-        return createHash("sha512").update(new Buffer(Math.floor(Math.random() * 1000))).digest("hex");
-    }
 }
 
 export {